Adapter Addition Continued

+++ app/src/main/java/com/example/moneymate/adapters/UserDetailAdapter.kt
@@ -2,7 +2,6 @@ package com.example.moneymate.adapters
 
 import android.app.AlertDialog
 import android.content.Context
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -13,7 +12,8 @@ import com.example.moneymate.R
 import com.example.moneymate.data.RegistrationTableHelper
 import com.example.moneymate.model.User
 
-internal class UserDetailAdapter(val context:Context):RecyclerView.Adapter<UserDetailAdapter.ViewHolder>(){
+internal class UserDetailAdapter(val context:Context) :
+    RecyclerView.Adapter<UserDetailAdapter.ViewHolder>() {
     private var usersList: ArrayList<User>
     init {
         val dbHelper = RegistrationTableHelper(context)
@@ -27,24 +27,21 @@ internal class UserDetailAdapter(val context:Context):RecyclerView.Adapter<UserD
             uidTextView.text = user.id.toString()
             val uEmailTextView: TextView = itemView.findViewById(R.id.ud_email)
             uEmailTextView.text = user.email
-            val userName: TextView = itemView.findViewById(R.id.ud_uname)
-            userName.text = user.uname
 
             btn.setOnClickListener {
-                val res = RegistrationTableHelper(context)
-                    .deleteUserById(user.id.toString())
-                if (res > 0) {
-                    val builder = AlertDialog.Builder(context)
-                    builder.setTitle("Delete ${user.uname} ?")
-                    builder.setMessage("Are you sure you want to Delete this user's account?")
-                    builder.setPositiveButton("Yes") { _, _ ->
+                val builder = AlertDialog.Builder(context)
+                builder.setTitle("Delete ${user.uname} ?")
+                builder.setMessage("Are you sure you want to Delete this user's account?")
+                builder.setPositiveButton("Yes") { _, _ ->
+                    val res = RegistrationTableHelper(context)
+                        .deleteUserById(user.id.toString())
+                    if (res > 0) {
                         usersList.removeAt(adapterPosition)
                         notifyItemRemoved(adapterPosition)
                     }
-                    builder.setNegativeButton("No", null)
-
-                    builder.show()
                 }
+                builder.setNegativeButton("No", null)
+                builder.show()
             }
         }
     }
@@ -56,7 +53,7 @@ internal class UserDetailAdapter(val context:Context):RecyclerView.Adapter<UserD
         return ViewHolder(myView)
     }
 
-    override fun getItemCount(): Int =usersList.size
+    override fun getItemCount(): Int = usersList.size
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         holder.setDetails(usersList[position])
